refactor: 더미데이터 패키지 이동
